It sort of works for orientation, it's "good" enough.

+++ Classes/FileViewController.h
@@ -28,8 +28,8 @@
 	// Image
 	UIScrollView *imageScrollView;
 	UIImageView *imageView;
-	int imageWidth;
-	int imageHeight;
+	float imageWidth;
+	float imageHeight;
 }
 
 @property (nonatomic, assign) id <FileViewDelegate> delegate;
